requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.73.0
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.121.2
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
huey==2.5.4
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightning==2.5.6
lightning-utilities==0.15.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==2.2.6
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-forecasting==1.5.0
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.49.3
statsmodels==0.14.5
sympy==1.14.0
threadpoolctl==3.6.0
-e git+ssh://git@github.com/rezaghasemi/ts_mini_project.git@f79c6965d43c510afe044402311cddce57dfbeb0#egg=time_series_forecasting
tomli==2.3.0
torch==2.9.1
torchmetrics==1.8.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
